+++ libraries/ibm_cert.rb
@@ -1,7 +1,8 @@
 #
 # Cookbook Name:: websphere
+# Resource:: ibm_cert
 #
-# Copyright (C) 2015 J Sainsburys
+# Copyright (C) 2015-2019 J Sainsburys
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -37,12 +38,13 @@ module WebsphereCookbook
 
     action :create do
       create_kdb
-      cmd = "#{ikeycmd} -cert -create -db #{kdb} -pw #{kdb_password} -sig_alg #{algorithm} -size #{size} "\
-        "-expire #{expire} -dn #{dn} -label #{label} -default_cert #{default_cert}"
+      cmd = "#{new_resource.ikeycmd} -cert -create -db #{new_resource.kdb} -pw #{new_resource.kdb_password}"\
+        " -sig_alg #{new_resource.algorithm} -size #{new_resource.size} -expire #{new_resource.expire}"\
+        " -dn #{new_resource.dn} -label #{new_resource.label} -default_cert #{new_resource.default_cert}"
 
-      execute "create cert #{label}" do
+      execute "create cert #{new_resource.label}" do
         command cmd
-        sensitive sensitive_exec
+        sensitive new_resource.sensitive_exec
         action :run
         returns [0, 22]
       end
@@ -51,65 +53,68 @@ module WebsphereCookbook
     end
 
     action :extract do
-      cmd = "#{ikeycmd} -cert -extract -db #{kdb} -pw #{kdb_password} -label #{label} -target #{extract_to} -format ascii"
+      cmd = "#{new_resource.ikeycmd} -cert -extract -db #{new_resource.kdb} -pw #{new_resource.kdb_password}"\
+        " -label #{new_resource.label} -target #{new_resource.extract_to} -format ascii"
 
-      execute "extract cert #{label}" do
+      execute "extract cert #{new_resource.label}" do
         command cmd
-        sensitive sensitive_exec
+        sensitive new_resource.sensitive_exec
         action :run
-        creates extract_to
-        only_if { ::File.exist?(kdb) }
+        creates new_resource.extract_to
+        only_if { ::File.exist?(new_resource.kdb) }
       end
 
       set_perms
     end
 
     action :set_default do
-      execute "set cert #{label} as default" do
-        command "#{ikeycmd} -cert -setdefault -pw #{kdb_password} -label #{label} -db #{kdb}"
-        sensitive sensitive_exec
+      execute "set cert #{new_resource.label} as default" do
+        command "#{new_resource.ikeycmd} -cert -setdefault -pw #{new_resource.kdb_password} -label #{new_resource.label} -db #{new_resource.kdb}"
+        sensitive new_resource.sensitive_exec
         action :run
         only_if { cert_in_keystore? }
       end
     end
 
     action :add do
-      execute "add cert #{label} to #{kdb}" do
-        command "#{ikeycmd} -cert -add -pw #{kdb_password} -label #{label} -trust enable -file #{add_cert} -db #{kdb}"
-        sensitive sensitive_exec
+      execute "add cert #{new_resource.label} to #{new_resource.kdb}" do
+        command "#{new_resource.ikeycmd} -cert -add -pw #{new_resource.kdb_password} -label #{new_resource.label} "\
+          "-trust enable -file #{new_resource.add_cert} -db #{new_resource.kdb}"
+        sensitive new_resource.sensitive_exec
         action :run
-        only_if { ::File.exist?(kdb) && ::File.exist?(add_cert) }
+        only_if { ::File.exist?(new_resource.kdb) && ::File.exist?(new_resource.add_cert) }
         not_if { cert_in_keystore? }
       end
     end
 
     action :import do
-      execute "import key database #{label} to #{kdb}" do
-        command "#{ikeycmd} -cert -import -target #{kdb} -target_pw #{kdb_password} -type #{kdb_type} -db #{add_cert} -label #{label} -target_type cms"
-        command << " -pw #{import_password}" if import_password
-        sensitive sensitive_exec
+      execute "import cert #{new_resource.label} to #{new_resource.kdb}" do
+        command "#{new_resource.ikeycmd} -cert -import -target #{new_resource.kdb} -target_pw #{new_resource.kdb_password} "\
+          "-type #{new_resource.kdb_type} -db #{new_resource.add_cert} -label #{new_resource.label} -target_type cms"
+        command << " -pw #{new_resource.import_password}" if new_resource.import_password
+        sensitive new_resource.sensitive_exec
         action :run
-        only_if { ::File.exist?(kdb) && ::File.exist?(add_cert) }
+        only_if { ::File.exist?(new_resource.kdb) && ::File.exist?(new_resource.add_cert) }
         not_if { cert_in_keystore? }
       end
     end
 
     action :update do
-      execute "remove matching key database #{label} from #{kdb}" do
-        command "#{ikeycmd} -cert -delete -db #{kdb} -pw #{kdb_password} -label #{label}"
-        sensitive sensitive_exec
+      execute "remove matching cert #{new_resource.label} from #{new_resource.kdb}" do
+        command "#{new_resource.ikeycmd} -cert -delete -db #{new_resource.kdb} -pw #{new_resource.kdb_password} -label #{new_resource.label}"
+        sensitive new_resource.sensitive_exec
         action :run
-        only_if { ::File.exist?(kdb) && ::File.exist?(add_cert) }
+        only_if { ::File.exist?(new_resource.kdb) && ::File.exist?(new_resource.add_cert) }
         not_if { matching_cert_in_keystore? }
-        notifies :run, "execute[import key database #{label} to #{kdb}]", :immediately
+        notifies :run, "execute[import cert #{new_resource.label} to #{new_resource.kdb}]", :immediately
       end
 
-      execute "import key database #{label} to #{kdb}" do
-        command "#{ikeycmd} -cert -import -target #{kdb} -target_pw #{kdb_password} -type #{kdb_type} -db #{add_cert} -label #{label} -target_type cms"
-        command << " -pw #{import_password}" if import_password
-        sensitive sensitive_exec
+      execute "import cert #{new_resource.label} to #{new_resource.kdb}" do
+        command "#{new_resource.ikeycmd} -cert -import -target #{new_resource.kdb} -target_pw #{new_resource.kdb_password} -type #{new_resource.kdb_type} -db #{new_resource.add_cert} -label #{new_resource.label} -target_type cms"
+        command << " -pw #{new_resource.import_password}" if new_resource.import_password
+        sensitive new_resource.sensitive_exec
         action :nothing
-        only_if { ::File.exist?(kdb) && ::File.exist?(add_cert) }
+        only_if { ::File.exist?(new_resource.kdb) && ::File.exist?(new_resource.add_cert) }
         not_if { cert_in_keystore? }
       end
     end
@@ -118,33 +123,33 @@ module WebsphereCookbook
     # so they're available in the action.
     action_class.class_eval do
       def create_kdb
-        dir = ::File.dirname(kdb)
+        dir = ::File.dirname(new_resource.kdb)
 
         directory dir do
           recursive true
-          mode 0600
-          owner owned_by
+          mode 0o600
+          owner new_resource.owned_by
           action :create
         end
 
-        execute "create kdb #{kdb}" do
-          command "#{ikeycmd} -keydb -create -db #{kdb} -pw #{kdb_password} -type cms -expire #{expire_kdb} -stash"
-          sensitive sensitive_exec
+        execute "create kdb #{new_resource.kdb}" do
+          command "#{new_resource.ikeycmd} -keydb -create -db #{new_resource.kdb} -pw #{new_resource.kdb_password} -type cms -expire #{new_resource.expire_kdb} -stash"
+          sensitive new_resource.sensitive_exec
           action :run
-          creates kdb
+          creates new_resource.kdb
         end
       end
 
       def set_perms
-        root_dir = ::File.dirname(kdb)
+        root_dir = ::File.dirname(new_resource.kdb)
         execute 'set perms' do
           cwd root_dir
-          command "chown #{owned_by} #{root_dir}/* && chmod 600 #{root_dir}/*"
+          command "chown #{new_resource.owned_by} #{root_dir}/* && chmod 600 #{root_dir}/*"
         end
       end
 
       def cert_sha256_fingerprint(db, cert_password = nil)
-        cmd = "#{ikeycmd} -cert -details -label #{label} -db #{db}"
+        cmd = "#{new_resource.ikeycmd} -cert -details -label #{new_resource.label} -db #{db}"
         cmd << " -pw #{cert_password}" if cert_password
         cmd << " | grep 'SHA256:' | awk '{print $2}'"
         mycmd = Mixlib::ShellOut.new(cmd, cwd: ::File.dirname(db))
@@ -153,14 +158,14 @@ module WebsphereCookbook
       end
 
       def cert_in_keystore?
-        cmd = "#{ikeycmd} -cert -list -pw #{kdb_password} -label #{label} -db #{kdb}"
-        mycmd = Mixlib::ShellOut.new(cmd, cwd: ::File.dirname(kdb))
+        cmd = "#{new_resource.ikeycmd} -cert -list -pw #{new_resource.kdb_password} -label #{new_resource.label} -db #{new_resource.kdb}"
+        mycmd = Mixlib::ShellOut.new(cmd, cwd: ::File.dirname(new_resource.kdb))
         mycmd.run_command
         if mycmd.stdout.include?("doesn't contain an entry with label") || mycmd.error?
-          Chef::Log.warn("certificate #{label} not found in #{kdb}")
+          Chef::Log.warn("certificate #{new_resource.label} not found in #{new_resource.kdb}")
           return false
         else
-          Chef::Log.warn("certificate #{label} already exists in #{kdb}")
+          Chef::Log.warn("certificate #{new_resource.label} already exists in #{new_resource.kdb}")
           return true
         end
       end
@@ -168,12 +173,12 @@ module WebsphereCookbook
       # return true if the SHA256 fingerprint matches the certificate in the keystore
       def matching_cert_in_keystore?
         if cert_in_keystore?
-          Chef::Log.warn("certificate #{label} found in #{kdb}, checking fingerprints")
-          kdb_print = cert_sha256_fingerprint(kdb, kdb_password)
-          cert_print = if import_password
-                         cert_sha256_fingerprint(add_cert, import_password)
+          Chef::Log.warn("certificate #{new_resource.label} found in #{new_resource.kdb}, checking fingerprints")
+          kdb_print = cert_sha256_fingerprint(new_resource.kdb, new_resource.kdb_password)
+          cert_print = if new_resource.import_password
+                         cert_sha256_fingerprint(new_resource.add_cert, new_resource.import_password)
                        else
-                         cert_sha256_fingerprint(add_cert)
+                         cert_sha256_fingerprint(new_resource.add_cert)
                        end
           return true if kdb_print == cert_print
         end