+++ libraries/websphere_jms_provider.rb
@@ -1,8 +1,8 @@
 #
 # Cookbook Name:: websphere
-# Resource:: websphere-server
+# Resource:: websphere_jms_provider
 #
-# Copyright (C) 2015 J Sainsburys
+# Copyright (C) 2015-2019 J Sainsburys
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -33,22 +33,22 @@ module WebsphereCookbook
     property :attributes, [Hash, nil], default: {} # Generic hash holding attributes for the provider - see : https://www.ibm.com/support/knowledgecenter/en/SS7K4U_7.0.0/com.ibm.websphere.zseries.doc/info/zseries/ae/rxml_7adminjms.html (ignore that it's for zOS)
     action :create do
       unless jms_provider_exists?
-        @jmsattrs = attributes.merge('classpath' => classpath_jars.join(';').to_s)
-        @jmsattrs['description'] = description
+        @jmsattrs = new_resource.attributes.merge('classpath' => new_resource.classpath_jars.join(';').to_s)
+        @jmsattrs['description'] = new_resource.description
         @attributes_str = attributes_to_wsadmin_str(@jmsattrs)
-        cmd = "AdminJMS.createJMSProviderAtScope('#{scope}', '#{provider_name}', "\
-          "'#{context_factory}', '#{url}', #{@attributes_str})"
+        cmd = "AdminJMS.createJMSProviderAtScope('#{new_resource.scope}', '#{new_resource.provider_name}', "\
+          "'#{new_resource.context_factory}', '#{new_resource.url}', #{@attributes_str})"
 
-        wsadmin_exec("Create JMS Provider #{provider_name}", cmd)
+        wsadmin_exec("Create JMS Provider #{new_resource.provider_name}", cmd)
       end
     end
     # need to wrap helper methods in class_eval
     # so they're available in the action.
     action_class.class_eval do
       def jms_provider_exists?
-        cmd = "-c \"AdminJMS.listJMSProviders('#{provider_name}')\""
+        cmd = "-c \"AdminJMS.listJMSProviders('#{new_resource.provider_name}')\""
         mycmd = wsadmin_returns(cmd)
-        return true if mycmd.stdout.include?("#{provider_name}\(")
+        return true if mycmd.stdout.include?("#{new_resource.provider_name}\(")
         false
       end
     end