+++ test/integration/was-cluster/default_spec.rb
@@ -1,12 +1,10 @@
-require 'spec_helper'
-
 describe group('ibm') do
   it { should exist }
 end
 
 describe user('ibm') do
   it { should exist }
-  it { belong_to_group 'ibm' }
+  its('groups') { should eq ['ibm'] }
 end
 
 describe file('/opt/IBM/InstallationManager/eclipse/tools/imcl') do
@@ -14,16 +12,15 @@ describe file('/opt/IBM/InstallationManager/eclipse/tools/imcl') do
 end
 
 describe command('/opt/IBM/InstallationManager/eclipse/tools/imcl listInstalledPackages') do
-  its(:stdout) { should match(/com.ibm.websphere.ND.v85_8.5.5010.*/) }
-  its(:stdout) { should match(/com.ibm.websphere.IBMJAVA.v71_7.1.*/) }
+  its(:stdout) { should match(/com.ibm.websphere.ND.v85_8.5.5014.*/) }
+  its(:stdout) { should match(/com.ibm.websphere.IBMJAVA.v80_8.0.*/) }
 end
 
 describe command('/opt/IBM/WebSphere/AppServer/bin/wsadmin.sh -lang jython -conntype SOAP -host localhost -user admin -password '\
   'admin -c "AdminServerManagement.listServers(\'\', \'\')"') do
-  its(:stdout) { should match(%r{nodeagent\(cells/MyNewCell/nodes/AppProfile1_node/servers/nodeagent}) }
   its(:stdout) { should match(%r{nodeagent\(cells/MyNewCell/nodes/CustomProfile1_node/servers/nodeagent}) }
-  its(:stdout) { should match(%r{ClusterServer1\(cells/MyNewCell/nodes/CustomProfile1_node/servers/ClusterServer1}) }
-  its(:stdout) { should match(%r{dmgr\(cells/MyNewCell/nodes/Dmgr01_node/servers/dmgr}) }
+  its(:stdout) { should match(%r{ClusterServer1-node\(cells/MyNewCell/nodes/CustomProfile1_node/servers/ClusterServer1}) }
+  its(:stdout) { should match(%r{dmgr\(cells/MyNewCell/nodes/MyNewNode/servers/dmgr}) }
 end
 
 # check dmgr security attribtue is set
@@ -32,7 +29,7 @@ describe command('/opt/IBM/WebSphere/AppServer/bin/wsadmin.sh -lang jython -conn
   its(:stdout) { should match(/true/) }
 end
 
-services = %w(Dmgr01 AppProfile1_node CustomProfile1_node)
+services = %w[Dmgr01 CustomProfile1_node]
 services.each do |service|
   describe service(service) do
     it { should be_enabled }
@@ -43,13 +40,13 @@ end
 describe file('/etc/init.d/Dmgr01') do
   its(:content) { should match(/startScript=startManager.sh/) }
 end
-describe file('/etc/init.d/AppProfile1_node') do
+describe file('/etc/init.d/CustomProfile1_node') do
   its(:content) { should match(/startScript=startServer.sh/) }
 end
 
-describe file('/opt/IBM/WebSphere/AppServer/profiles/Dmgr01/config/cells/MyNewCell/nodes/CustomProfile1_node/servers/ClusterServer1/server.xml') do
-  it { should be_owned_by 'was' }
-  it { should be_grouped_into 'was' }
+describe file('/opt/IBM/WebSphere/AppServer/profiles/Dmgr01/config/cells/MyNewCell/nodes/CustomProfile1_node/servers/ClusterServer1-node/server.xml') do
+  its('owner') { should eq 'was' }
+  its('group') { should eq 'was' }
   its(:content) { should match(/<jvmEntries.*initialHeapSize="1024".*maximumHeapSize="3072".*debugMode="true"/) }
   its(:content) { should match(/maximumStartupAttempts="2"/) }
   its(:content) { should match(/pingInterval="222"/) }
@@ -59,16 +56,20 @@ describe file('/opt/IBM/WebSphere/AppServer/profiles/Dmgr01/config/cells/MyNewCe
   its(:content) { should match(/systemProperties.*name="my_system_property" value="testing123"/) }
 end
 
-ports = %w(9043 8879)
+ports = %w[9043 8879]
 ports.each do |p|
   describe port(p) do
     it { should be_listening }
   end
 end
 
+describe processes(Regexp.new('/opt/IBM/WebSphere/AppServer/java/bin/java.*MyNewCell CustomProfile1_node nodeagent$')) do
+  it { should exist }
+end
+
 # java tests
 describe command('/opt/IBM/WebSphere/AppServer/bin/managesdk.sh -listEnabledProfile -profileName CustomProfile1') do
-  its(:stdout) { should match(/Server ClusterServer1 SDK name: 1.7.1_64/) }
+  its(:stdout) { should match(/Server ClusterServer1-node SDK name: 1.8_64/) }
 end
 
 # cluster tests